+++ all.sh
@@ -13,11 +13,11 @@
 # functionality by running each modules driver file 
 
 # build + run test drivers in Makefile 
-make arith
-make calculus
-make linalg
-make ml_dl
-make num-theory
+make arith -j4
+make calculus -j4
+make linalg -j4
+make ml_dl -j4
+make num-theory -j4
 
 # delete generated binaries
 make clean-mods
