+++ drivers/ml/d_reg.cpp
@@ -3,7 +3,7 @@
  */
 #include <iostream>
 #include <stdio.h>
-#include "../../include/ML_DL/linreg.hpp"
+#include "../../include/ml/linreg.hpp"
 
 
 int main() {
