+++ include/ml/kohonen_network.hpp
