+++ modules/ml/naive_network.cpp
@@ -2,7 +2,7 @@
  * Implementation of a standard or naive neural network
  */
 #include "../../include/linalg/matrices.hpp"
-#include "../../include/ML_DL/naive_network.hpp"
+#include "../../include/ml/naive_network.hpp"
